feat(acl:parser): ACLCanType转ACLType转换器，修复bug #952
